requirements.txt
fastapi
uvicorn
sqlalchemy  
passlib
python-jose
python-multipart
pydantic
pymysql
cryptography
